changed_article_moved_modal_and_form
