+++ src/App.css
@@ -31,3 +31,19 @@
   position: absolute;
   width: 100%;
 }
+
+.apparrow {
+  border: solid black;
+  border-width: 0 3px 3px 0;
+  display: inline-block;
+  padding: 3px;
+  width: 15px;
+  height: 15px;
+  position: absolute;
+  right: 5px;
+  top: 320px;
+}
+.appright {
+  transform: rotate(-45deg);
+  -webkit-transform: rotate(-45deg);
+}
