+++ src/components/articles/Articles.js
@@ -5,24 +5,8 @@ import {
   Accordion, AccordionItem, AccordionItemTitle,AccordionItemBody} from "react-accessible-accordion";
 import "react-accessible-accordion/dist/fancy-example.css";
 import ReplaceSource from '../ReplaceSource'
+import Modal from '../modal/Modal'
 import './article.css'
-import Modal from 'react-modal'
-
-const modalStyles = {
-  content:{
-    top                   : '50%',
-    left                  : '50%',
-    right                 : 'auto',
-    bottom                : 'auto',
-    marginRight           : '-50%',
-    transform             : 'translate(-50%, -50%)',
-    width                 : '90%',
-    height                : '80%',
-    borderRadius          : '10px',
-    boxShadow             : '0px 5px 3px 2px rgba(190,200,200,0.5)',
-    background            : 'linear-gradient(red,orange,yellow,blue)'
-  }
-}
 
 
 class Articles extends Component {
@@ -56,6 +40,7 @@ class Articles extends Component {
 
     return (
       <div className='accordian_style'>
+        <i className='arrow right' onClick={()=>this.openModal()}></i>
         <Accordion>
           <AccordionItem>
             <AccordionItemTitle>
@@ -76,35 +61,12 @@ class Articles extends Component {
             </AccordionItemBody>
           </AccordionItem>
         </Accordion>
-        <i className='arrow right' onClick={()=>this.openModal()}></i>
-        <Modal
-        isOpen={this.state.modal}
-        onRequestClose={()=>this.closeModal()}
-        style={modalStyles}
-        >
-        <form className='modal'>
-          <p>Which article do you feel was More Reasonable</p>
-          <label htmlFor="article1">Article 1</label>
-          <input name='article' value='article1' type="radio"/>
-          <label htmlFor="article2">Article 2</label>
-          <input name='article' value='article2' type="radio"/>
-        </form>
-        <form className='modal two'>
-            <p>Surprised</p>
-            <label htmlFor="yes">Yes</label>
-            <input name='yesorno' value='yes' type="radio"/>
-            <label htmlFor="no">No</label>
-            <input name='yesorno' value='no' type="radio"/>
-            <button onClick={()=>this.closeModal()}>Close/Next</button>
-        </form>
-        </Modal>
-        <Modal
-        isOpen={this.state.graphModal}
-        onRequestClose={()=>this.closeGraphModal()}
-        style={modalStyles}
-        >
-        <h1>Graph Modal</h1>
-        </Modal>
+        <Modal 
+          modal={this.state.modal} 
+          closeModal={()=>this.closeModal()}
+          graphModal={this.state.graphModal}
+          closeGraphModal={()=>this.closeGraphModal()}
+        />
       </div>
     );
   }
