+++ src/components/intro/intro.css
@@ -30,3 +30,5 @@ section {  /* also used by Topics.css */
     text-decoration: none;
     padding:4vw; 
 }
+
+
